pyproject.toml
[project]
name = "online-cinema-api"
version = "0.1.0"
description = ""
authors = [
    {name = "meh84",email = "meh8474@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-stubs (>=5.2.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "stripe (>=12.3.0,<13.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

